import newlib-2000-02-17 snapshot

+++ newlib/libm/mathfp/sf_sqrt.c
@@ -94,7 +94,7 @@ _DEFUN (sqrtf, (float),
 
 double sqrt (double x)
 {
-  return (double) sqrtf ((float) x);
+  return (double) sqrt ((float) x);
 }
 
 #endif /* _DOUBLE_IS_32BITS */
